[import_ttf] -Fix: do the approximation forcing doubles
git-svn-id: svn://repo.hu/pcb-rnd/trunk@20264 1d61de99-560f-4dba-8957-1e505d42358b

+++ src_plugins/import_ttf/str_approx.c
@@ -47,8 +47,8 @@ int stroke_approx_conic_to(const FT_Vector *control, const FT_Vector *to, void *
 
 	if (str_approx_comment != NULL) printf("%s conic to {\n", str_approx_comment);
 	for(t = 0.0; t <= 1.0; t += td) {
-		v.x = sqr(1.0-t) * s->x + 2*t*(1.0-t)*control->x + t*t*to->x;
-		v.y = sqr(1.0-t) * s->y + 2*t*(1.0-t)*control->y + t*t*to->y;
+		v.x = sqr(1.0-t) * s->x + 2.0*t*(1.0-t)*(double)control->x + t*t*(double)to->x;
+		v.y = sqr(1.0-t) * s->y + 2.0*t*(1.0-t)*(double)control->y + t*t*(double)to->y;
 		s->funcs.line_to(&v, s);
 	}
 	if (str_approx_comment != NULL) printf("%s }\n", str_approx_comment);
